Move parser integration tests into an integration folder
