+++ tests/test_parser.py
@@ -2,27 +2,6 @@ import pytest
 from .setup_transaction_tests import chain as s, tester as t, ethereum_utils as u, check_gas, \
     get_contract_with_gas_estimation, get_contract
 
-def test_null_code():
-    null_code = """
-def foo():
-    pass
-    """
-    c = get_contract_with_gas_estimation(null_code)
-    c.foo()
-    print('Successfully executed a null function')
-
-
-def test_basic_code():
-    basic_code = """
-
-def foo(x: num) -> num:
-    return x * 2
-
-    """
-    c = get_contract_with_gas_estimation(basic_code)
-    assert c.foo(9) == 18
-    print('Passed basic code test')
-
 
 def test_basic_repeater():
     basic_repeater = """
@@ -122,75 +101,6 @@ def foo() -> num:
     c.oo()
     assert c.foo() == 5
     print('Passed basic state accessor test')
-
-
-def test_arbitration_code():
-    arbitration_code = """
-buyer: address
-seller: address
-arbitrator: address
-
-def setup(_seller: address, _arbitrator: address):
-    if not self.buyer:
-        self.buyer = msg.sender
-        self.seller = _seller
-        self.arbitrator = _arbitrator
-
-def finalize():
-    assert msg.sender == self.buyer or msg.sender == self.arbitrator
-    send(self.seller, self.balance)
-
-def refund():
-    assert msg.sender == self.seller or msg.sender == self.arbitrator
-    send(self.buyer, self.balance)
-
-    """
-
-    c = get_contract_with_gas_estimation(arbitration_code, value=1)
-    c.setup(t.a1, t.a2, sender=t.k0)
-    try:
-        c.finalize(sender=t.k1)
-        success = True
-    except:
-        success = False
-    assert not success
-    c.finalize(sender=t.k0)
-
-    print('Passed escrow test')
-
-
-def test_arbitration_code_with_init():
-    arbitration_code_with_init = """
-buyer: address
-seller: address
-arbitrator: address
-
-@payable
-def __init__(_seller: address, _arbitrator: address):
-    if not self.buyer:
-        self.buyer = msg.sender
-        self.seller = _seller
-        self.arbitrator = _arbitrator
-
-def finalize():
-    assert msg.sender == self.buyer or msg.sender == self.arbitrator
-    send(self.seller, self.balance)
-
-def refund():
-    assert msg.sender == self.seller or msg.sender == self.arbitrator
-    send(self.buyer, self.balance)
-    """
-
-    c = get_contract_with_gas_estimation(arbitration_code_with_init, args=[t.a1, t.a2], sender=t.k0, value=1)
-    try:
-        c.finalize(sender=t.k1)
-        success = True
-    except t.TransactionFailed:
-        success = False
-    assert not success
-    c.finalize(sender=t.k0)
-
-    print('Passed escrow test with initializer')
 
 
 def test_send():
@@ -376,100 +286,6 @@ def returnMoose() -> num:
     c = get_contract_with_gas_estimation(permanent_variables_test, args=[5, 7])
     assert c.returnMoose() == 57
     print('Passed init argument and variable member test')
-
-
-def test_crowdfund():
-    crowdfund = """
-
-funders: {sender: address, value: wei_value}[num]
-nextFunderIndex: num
-beneficiary: address
-deadline: timestamp
-goal: wei_value
-refundIndex: num
-timelimit: timedelta
-
-def __init__(_beneficiary: address, _goal: wei_value, _timelimit: timedelta):
-    self.beneficiary = _beneficiary
-    self.deadline = block.timestamp + _timelimit
-    self.timelimit = _timelimit
-    self.goal = _goal
-
-@payable
-def participate():
-    assert block.timestamp < self.deadline
-    nfi = self.nextFunderIndex
-    self.funders[nfi].sender = msg.sender
-    self.funders[nfi].value = msg.value
-    self.nextFunderIndex = nfi + 1
-
-@constant
-def expired() -> bool:
-    return block.timestamp >= self.deadline
-
-@constant
-def timestamp() -> timestamp:
-    return block.timestamp
-
-@constant
-def deadline() -> timestamp:
-    return self.deadline
-
-@constant
-def timelimit() -> timedelta:
-    return self.timelimit
-
-@constant
-def reached() -> bool:
-    return self.balance >= self.goal
-
-def finalize():
-    assert block.timestamp >= self.deadline and self.balance >= self.goal
-    selfdestruct(self.beneficiary)
-
-def refund():
-    ind = self.refundIndex
-    for i in range(ind, ind + 30):
-        if i >= self.nextFunderIndex:
-            self.refundIndex = self.nextFunderIndex
-            return
-        send(self.funders[i].sender, self.funders[i].value)
-        self.funders[i].sender = 0x0000000000000000000000000000000000000000
-        self.funders[i].value = 0
-    self.refundIndex = ind + 30
-
-    """
-
-    c = get_contract(crowdfund, args=[t.a1, 50, 600])
-
-    c.participate(value=5)
-    assert c.timelimit() == 600
-    assert c.deadline() - c.timestamp() == 600
-    assert not c.expired()
-    assert not c.reached()
-    c.participate(value=49)
-    assert c.reached()
-    pre_bal = s.head_state.get_balance(t.a1)
-    s.head_state.timestamp += 1000
-    assert c.expired()
-    c.finalize()
-    post_bal = s.head_state.get_balance(t.a1)
-    assert post_bal - pre_bal == 54
-
-    c = get_contract(crowdfund, args=[t.a1, 50, 600])
-    c.participate(value=1, sender=t.k3)
-    c.participate(value=2, sender=t.k4)
-    c.participate(value=3, sender=t.k5)
-    c.participate(value=4, sender=t.k6)
-    s.head_state.timestamp += 1000
-    assert c.expired()
-    assert not c.reached()
-    pre_bals = [s.head_state.get_balance(x) for x in [t.a3, t.a4, t.a5, t.a6]]
-    c.refund()
-    post_bals = [s.head_state.get_balance(x) for x in [t.a3, t.a4, t.a5, t.a6]]
-    assert [y-x for x, y in zip(pre_bals, post_bals)] == [1, 2, 3, 4]
-
-    print('Passed composite crowdfund test')
 
 
 def test_comment_test():
@@ -696,98 +512,6 @@ def foq() -> num:
     assert c.fop() == 4625
     assert c.foq() == 4020
     print('Passed composite struct test')
-
-
-def test_crowdfund2():
-    crowdfund2 = """
-
-funders: {sender: address, value: wei_value}[num]
-nextFunderIndex: num
-beneficiary: address
-deadline: timestamp
-goal: wei_value
-refundIndex: num
-timelimit: timedelta
-
-def __init__(_beneficiary: address, _goal: wei_value, _timelimit: timedelta):
-    self.beneficiary = _beneficiary
-    self.deadline = block.timestamp + _timelimit
-    self.timelimit = _timelimit
-    self.goal = _goal
-
-@payable
-def participate():
-    assert block.timestamp < self.deadline
-    nfi = self.nextFunderIndex
-    self.funders[nfi] = {sender: msg.sender, value: msg.value}
-    self.nextFunderIndex = nfi + 1
-
-@constant
-def expired() -> bool:
-    return block.timestamp >= self.deadline
-
-@constant
-def timestamp() -> timestamp:
-    return block.timestamp
-
-@constant
-def deadline() -> timestamp:
-    return self.deadline
-
-@constant
-def timelimit() -> timedelta:
-    return self.timelimit
-
-@constant
-def reached() -> bool:
-    return self.balance >= self.goal
-
-def finalize():
-    assert block.timestamp >= self.deadline and self.balance >= self.goal
-    selfdestruct(self.beneficiary)
-
-def refund():
-    ind = self.refundIndex
-    for i in range(ind, ind + 30):
-        if i >= self.nextFunderIndex:
-            self.refundIndex = self.nextFunderIndex
-            return
-        send(self.funders[i].sender, self.funders[i].value)
-        self.funders[i] = None
-    self.refundIndex = ind + 30
-
-    """
-
-    c = get_contract(crowdfund2, args=[t.a1, 50, 600])
-
-    c.participate(value=5)
-    assert c.timelimit() == 600
-    assert c.deadline() - c.timestamp() == 600
-    assert not c.expired()
-    assert not c.reached()
-    c.participate(value=49)
-    assert c.reached()
-    pre_bal = s.head_state.get_balance(t.a1)
-    s.head_state.timestamp += 1000
-    assert c.expired()
-    c.finalize()
-    post_bal = s.head_state.get_balance(t.a1)
-    assert post_bal - pre_bal == 54
-
-    c = get_contract(crowdfund2, args=[t.a1, 50, 600])
-    c.participate(value=1, sender=t.k3)
-    c.participate(value=2, sender=t.k4)
-    c.participate(value=3, sender=t.k5)
-    c.participate(value=4, sender=t.k6)
-    s.head_state.timestamp += 1000
-    assert c.expired()
-    assert not c.reached()
-    pre_bals = [s.head_state.get_balance(x) for x in [t.a3, t.a4, t.a5, t.a6]]
-    c.refund()
-    post_bals = [s.head_state.get_balance(x) for x in [t.a3, t.a4, t.a5, t.a6]]
-    assert [y-x for x, y in zip(pre_bals, post_bals)] == [1, 2, 3, 4]
-
-    print('Passed second composite crowdfund test')
 
 
 def test_test_slice():
